Add GCS remote backend
