+++ go.mod
@@ -3,14 +3,16 @@ module github.com/fujiwara/tfstate-lookup
 go 1.14
 
 require (
+	cloud.google.com/go/storage v1.15.0
 	github.com/aws/aws-sdk-go v1.30.7
 	github.com/fastly/go-utils v0.0.0-20180712184237-d95a45783239 // indirect
-	github.com/google/go-cmp v0.3.1
+	github.com/google/go-cmp v0.5.5
 	github.com/itchyny/gojq v0.9.0
 	github.com/jehiah/go-strftime v0.0.0-20171201141054-1d33003b3869 // indirect
 	github.com/kr/pretty v0.1.0 // indirect
 	github.com/mattn/go-isatty v0.0.12
 	github.com/pkg/errors v0.9.1
 	github.com/tebeka/strftime v0.1.3 // indirect
+	google.golang.org/api v0.45.0
 	gopkg.in/check.v1 v1.0.0-20180628173108-788fd7840127 // indirect
 )
